[Fix #359] 방 입장 시 players의 초기 Ready 상태인 "userInfo" socketEvent 위치 변경
