+++ components/mafia/MafiaPlayRooms.tsx
@@ -6,19 +6,22 @@ import { useGameActions, useGameState } from "@/store/game-store";
 import { useOverLayActions } from "@/store/overlay-store";
 import { useModalActions } from "@/store/show-modal-store";
 import S from "@/style/livekit/livekit.module.css";
-import { MediaStatus } from "@/types";
+import { MediaStatus, playersInfo } from "@/types";
 import { allAudioSetting } from "@/utils/participantCamSettings/camSetting";
 import { RoomAudioRenderer, useLocalParticipant, useTracks } from "@livekit/components-react";
 import { Track } from "livekit-client";
-import { useEffect } from "react";
+import { useEffect, useRef } from "react";
 import MafiaHeader from "./MafiaHeader";
 import LocalParticipant from "./LocalParticipant";
 import MafiaModals from "./MafiaModals";
 import MafiaToolTip from "./MafiaToolTip";
 import RemoteParticipant from "./RemoteParticipant";
+import { socket } from "@/utils/socket/socket";
 
 const MafiaPlayRooms = () => {
-  const { localParticipant } = useLocalParticipant();
+  const hasEmitted = useRef(false);
+
+  const localParticipant = useLocalParticipant();
   const isGameState = useGameState();
   const { setDiedPlayer, setIsGameState, setGameReset } = useGameActions();
   const { setReadyPlayers, setOverlayReset } = useOverLayActions();
@@ -26,21 +29,22 @@ const MafiaPlayRooms = () => {
   const { setIsMediaReset, setPlayersMediaStatus } = useMediaDevice(); // 카메라 및 오디오 처리
   useSelectSocket(); // 클릭 이벤트 처리
 
-  //NOTE -  전체 데이터
-  const tracks = useTracks(
-    [
-      { source: Track.Source.Camera, withPlaceholder: true },
-      { source: Track.Source.Microphone, withPlaceholder: true }
-    ],
-    { onlySubscribed: true } // 구독됐을 경우에만 실행
-  );
-
   //NOTE - 방 입장 시 초기화
   useEffect(() => {
     setOverlayReset(); //Local,Remote 클릭 이벤트 및 캠 이미지 초기화
     setModalReset(); //전체 모달 요소 초기화
     setGameReset(); // 죽은 players 및 게임 state 초기화
   }, []);
+
+  //NOTE - 방 입장 시 방의 정보를 불러온다.
+  useEffect(() => {
+    if (localParticipant.localParticipant.metadata && !hasEmitted.current) {
+      const roomId = localParticipant.localParticipant.metadata;
+
+      socket.emit("usersInfo", roomId);
+      hasEmitted.current = true;
+    }
+  }, [localParticipant]);
 
   const sockets = {
     //NOTE - 전체 players의 실시간 Ready 상태
@@ -63,6 +67,12 @@ const MafiaPlayRooms = () => {
     //NOTE - 방의 변화를 감지
     updateRoomInfo: (roomInfo: any) => {
       console.log("roomInfo", roomInfo);
+    },
+    //NOTE - players의 초기 Ready 상태
+    usersInfo: (players: playersInfo[]) => {
+      players.forEach((player) => {
+        setReadyPlayers(player.user_id, player.is_ready);
+      });
     },
     //NOTE - Error 처리
     playError: (roomName: string, error: string) => {
@@ -92,18 +102,18 @@ const MafiaPlayRooms = () => {
   return (
     <section className={`${S.mafiaPlayRoomWrapper} ${pretendard.className}`}>
       <div className={S.goToMainPage}></div>
-      <button
+      {/* <button
         onClick={() => {
           allAudioSetting(tracks, false);
         }}
         style={{ background: "red" }}
       >
         전체 소리 끄기
-      </button>
+      </button> */}
       <MafiaHeader />
       <div className={S.mafiaPlayRoomSection}>
-        <LocalParticipant tracks={tracks} />
-        <RemoteParticipant tracks={tracks} />
+        <LocalParticipant />
+        <RemoteParticipant />
         <RoomAudioRenderer muted={false} /> {/* 원격 참가자 오디오 트랙 처리 및 관리 */}
         <MafiaToolTip />
         <MafiaModals />
